% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofellcp.R
\docType{package}
\name{gofellcp}
\alias{gofellcp}
\title{gofellcp: Goodness-of-fit tests for elliptical copulas}
\description{
The gofellcp package provides the tests:
testEquality2D, testEqualityMult, testEqualityMult_L2subsam,
testSymmetry, testRadialSymmetry, testProEllip
}
